refactor(telegram-bot): split TelegramBotService into focused collaborators (#715)
The 1,882-line TelegramBotService god class is decomposed into a thin
lifecycle/transport shell plus single-responsibility collaborators, with
the embedded model maps and duplicated logic extracted to one place each.
Behavior is fully preserved (identical commands, replies, routing).

New modules (all <=~380 lines, methods <=~120):
- telegram-bot.types.ts            shared types
- telegram-bot-model-maps.constants.ts  fal/Replicate model maps (out of initializeEngine)
- telegram-command-args.util.ts    extractCommandArgs helper
- telegram-workflow-executors.ts   registerWorkflowExecutors + pollReplicatePrediction;
                                   collapses the identical passthrough executors
- telegram-workflow-loader.ts      load/extractInputs/toExecutableWorkflow;
                                   single node-type -> input-kind mapping
- telegram-run-commands.service.ts run control-plane + auth context + /connect/.../status<id>
- telegram-workflow-runner.service.ts  execution + progress + results
- telegram-conversation.service.ts conversation state machine (sole state owner)
- telegram-message-handler.service.ts  photo/text intake + photo download

Dedup: passthrough executors, node-type mapping, pending-image fan-out, and
the cancel/reset block now each exist in exactly one place.

Tests: existing 23 specs unchanged and green; added unit tests for the
extracted executors, loader, and conversation service (38 total).
